requirements-benchmark.txt
# Core cryptographic libraries
charm-crypto>=0.50
pycrypto>=2.6.1

# MA-ABE and pairing-based cryptography
pypbc>=0.1.0

# Symmetric encryption
cryptocode>=0.1

# IPFS client
ipfshttpclient>=0.8.0a2

# System monitoring
psutil>=5.8.0

# Data visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Data processing and analysis
pandas>=1.3.0
numpy>=1.21.0

# Configuration management
python-decouple>=3.4

# Database
sqlite3

# Standard libraries (usually included but listed for completeness)
statistics
json
datetime
tempfile
subprocess
hashlib
random
time
os
sys

# Additional utilities
tqdm>=4.62.0  # Progress bars
tabulate>=0.8.9  # Pretty tables 